requirements.txt
Flask
Flask-Cors
moviepy
python-dotenv
Werkzeug